ajuste de estructura
