+++ frontend/src/pages/dashboard.css
@@ -1,0 +1,115 @@
+.stat-card {
+    border-radius: 1rem;
+    box-shadow: 0 8px 30px rgba(0,0,0,0.1);
+    transition: all 0.3s ease;
+    border: none;
+    position: relative;
+    z-index: 2;
+}
+  
+.stat-card:hover {
+    transform: translateY(-5px);
+    box-shadow: 0 15px 40px rgba(0,0,0,0.15);
+  }
+  
+.stat-card.clickable {
+    cursor: pointer;
+}
+  
+.action-card {
+    border: 2px solid #e9ecef;
+    border-radius: 0.75rem;
+    transition: all 0.3s ease;
+}
+  
+.action-card:hover {
+    box-shadow: 0 10px 25px rgba(0,0,0,0.1);
+    border-color: rgba(220, 53, 69, 0.3);
+    transform: translateY(-2px);
+}
+  
+.queue-item {
+    border-left: 4px solid;
+    border-radius: 0 0.5rem 0.5rem 0;
+}
+  
+.priority-high {
+    border-left-color: #dc3545;
+}
+  
+.priority-normal {
+    border-left-color: #fd7e14;
+}
+    
+/* Estilos para el turno en atención */
+.current-turn-card {
+    background: linear-gradient(135deg, #fd0d0dff 0%, #750b0bff 100%);
+    color: white;
+    border-radius: 1rem;
+    padding: 1rem;
+    position: relative;
+    overflow: hidden;
+}
+  
+.current-turn-card::before {
+    content: '';
+    position: absolute;
+    top: -40%;
+    right: -50%;
+    width: 100%;
+    height: 200%;
+    background: rgba(255, 255, 255, 0.1);
+    transform: rotate(45deg);
+    z-index: 1;
+}
+  
+.current-turn-badge {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    background: rgba(255, 255, 255, 0.2);
+    padding: 0.5rem 1rem;
+    border-radius: 2rem;
+    font-weight: bold;
+    z-index: 2;
+}
+  
+.turn-number-display {
+    font-size: 4rem;
+    font-weight: bold;
+    margin-bottom: 0.5rem;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+    z-index: 2;
+    position: relative;
+}
+  
+.customer-name {
+    font-size: 1.8rem;
+    margin-bottom: 0.5rem;
+    z-index: 2;
+    position: relative;
+}
+  
+.turn-details {
+    background: rgba(255, 255, 255, 0.15);
+    border-radius: 0.75rem;
+    padding: 1rem;
+    margin-top: 1.5rem;
+    z-index: 2;
+    position: relative;
+}
+  
+.detail-item {
+    display: flex;
+    align-items: center;
+    margin-bottom: 0.5rem;
+}
+  
+.detail-item:last-child {
+    margin-bottom: 0;
+}
+  
+.detail-icon {
+    margin-right: 0.75rem;
+    width: 20px;
+}
